+++ protocol/protocol.go
@@ -25,6 +25,7 @@ import (
 	"code.vegaprotocol.io/vega/logging"
 	"code.vegaprotocol.io/vega/nodewallets"
 	"code.vegaprotocol.io/vega/processor"
+	"code.vegaprotocol.io/vega/protocolupgrade"
 	"code.vegaprotocol.io/vega/stats"
 	"code.vegaprotocol.io/vega/subscribers"
 	"code.vegaprotocol.io/vega/vegatime"
@@ -111,6 +112,7 @@ func New(
 			svcs.blockchainClient,
 			svcs.erc20MultiSigTopology,
 			stats.GetVersion(),
+			svcs.protocolUpgradeEngine,
 			svcs.codec,
 		),
 		log:         log,
@@ -156,4 +158,8 @@ func (n *Protocol) GetBroker() *broker.Broker {
 
 func (n *Protocol) GetPoW() api.ProofOfWorkParams {
 	return n.services.pow
+}
+
+func (n *Protocol) GetProtocolUpgradeService() *protocolupgrade.Engine {
+	return n.services.protocolUpgradeEngine
 }