+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The init command now also generate the configuration for tendermint, the flags `
 ### 🛠 Improvements
 - [5428](https://github.com/vegaprotocol/vega/pull/5428) - Update contributor information
 - [5519](https://github.com/vegaprotocol/vega/pull/5519) - Add `--genesis-file` option to the `load_checkpoint` command
+- [5538](https://github.com/vegaprotocol/vega/issues/5538) - Core side implementation of protocol upgrade
 - [5525](https://github.com/vegaprotocol/vega/pull/5525) - Release `vegawallet` from the core
 - [5524](https://github.com/vegaprotocol/vega/pull/5524) - Align `vegawallet` and core versions
 - [5524](https://github.com/vegaprotocol/vega/pull/5549) - Add endpoint for getting the network's `chain-id`